Added reveiw model and views

+++ brainstorm/payment/models.py
@@ -83,7 +83,7 @@ class OrderItem(models.Model):
     product = models.ForeignKey(Product, on_delete=models.CASCADE, null=True)
 
 
-    quantity = models.PositiveBigIntegerField(default=1)
+    quantity = models.PositiveIntegerField(default=1)
 
     price = models.DecimalField(max_digits=8, decimal_places=2)
